fix: compare 오류 수정 중
