+++ pages/compare/index.tsx
@@ -12,210 +12,38 @@ import { CompareChipA, CompareChipB } from "@/src/components/common/chip/Compare
 import { useQuery } from "@tanstack/react-query";
 import { QUERY_KEY } from "@/src/routes";
 import { getProducts } from "@/src/apis/product";
-import styled, { css } from "styled-components";
-import { fontStyle } from "@/styles/theme";
 import { CompareTable } from "@/src/components/compare/Compare";
+import Product1 from "@/src/components/compare/Product1";
+import Product2 from "@/src/components/compare/Product2";
 
 type CompareProps = {
   $productName: string;
 };
 
-export default function Compare({ $productName }: CompareProps) {
-  const [product1, setProduct1] = useState("");
-  const [product1Chip, setProduct1Chip] = useState("");
-  const [readOnly, setReadOnly] = useState(false);
-  // const [readOnly2, setReadOnly2] = useState(false);
-  const [isShowChip1, setIsShowChip1] = useState(false);
-  // const [product2, setProduct2] = useState("");
-  // const [product2Chip, setProduct2Chip] = useState("");
-  // const [isShowChip2, setIsShowChip2] = useState(false);
-  const [isShow, setIsShow] = useState(false);
-  // const [isShow2, setIsShow2] = useState(false);
-  // const [newProduct1, setNewProduct1] = useState("");
-  // const [newProduct2, setNewProduct2] = useState("");
-
-  const handleProductAChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setProduct1(e.target.value);
-    setProduct1Chip(e.target.value);
-    setReadOnly(false);
-    if (e.target.value === "") {
-      handleDelete();
-    }
-  };
-  // const handleProductBChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-  //   setProduct2(e.target.value);
-  //   setProduct2Chip(e.target.value);
-  //   setReadOnly2(false);
-  //   if (e.target.value === "") {
-  //     handleDelete();
-  //   }
-  // };
-
-  const handleAddProductA = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === "Enter" && product1.trim() !== "") {
-      setProduct1Chip(e.target.value);
-      setProduct1("");
-      setReadOnly(true);
-      setIsShowChip1(true);
-    }
-  };
-
-  // const handleAddProductB = (e: React.KeyboardEvent<HTMLInputElement>) => {
-  //   if (e.key === "Enter" && product1.trim() !== "") {
-  //     setProduct2Chip(e.target.value);
-  //     setProduct2("");
-  //     setReadOnly2(true);
-  //     setIsShowChip2(true);
-  //   }
-  // };
-
-  const handleDelete = () => {
-    setIsShowChip1(false);
-    setReadOnly(false);
-  };
-  const handleClickSearch = (e: React.MouseEvent<HTMLInputElement>) => {
-    const productName = e.currentTarget.innerText;
-    setProduct1Chip(productName);
-    setReadOnly(true);
-    setIsShowChip1(true);
-    setProduct1("");
-  };
-  // const handleClickSearch2 = (e: React.MouseEvent<HTMLInputElement>) => {
-  //   const productName = e.currentTarget.innerText;
-  //   setProduct2Chip(productName);
-  //   setReadOnly2(true);
-  //   setIsShowChip2(true);
-  //   setProduct2("");
-  // };
-
-  const { data: similarProducts, isSuccess } = useQuery({
-    queryKey: [QUERY_KEY.PRODUCTS, { keyword: product1Chip }],
-    queryFn: () => getProducts({ keyword: product1Chip }),
-    enabled: !!product1Chip,
-  });
-
-  // const { data: similarProducts2, isSuccess } = useQuery({
-  //   queryKey: [QUERY_KEY.PRODUCTS, { keyword: product2Chip }],
-  //   queryFn: () => getProducts({ keyword: product2Chip }),
-  //   enabled: !!product2Chip,
-  // });
-
+export default function Compare(): CompareProps {
   const [productOneData, setProductOneData] = useState<any>();
-  // const [productTwoData, setProductTwoData] = useState<any>();
+  const [similarProducts, setSimilarProducts] = useState<any>(); // similarProducts 상태 추가
 
-  useEffect(() => {
-    setIsShow(false);
-    console.log(product1);
-    if (isSuccess) {
-      if (product1 === "") {
-        setIsShow(false);
-      } else {
-        setIsShow(true);
-      }
-    }
-  }, [isSuccess, similarProducts, product1Chip, productOneData]);
-
-  const handleSubmit = (event): any => {
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
     setProductOneData(similarProducts?.list[0]);
   };
-
+  const handleSearchProduct = async (productName: string) => {
+    const data = await getProducts({ keyword: productName });
+    setSimilarProducts(data);
+  };
   return (
     <>
       <BodyContainer>
-        <InputContainer>
-          <Label>상품1</Label>
-          <TagContainer>
-            <ProductInput
-              value={product1}
-              onChange={handleProductAChange}
-              onKeyDown={handleAddProductA}
-              readOnly={readOnly}
-            />
-            {isShowChip1 && (
-              <ChipContainer onClick={handleDelete}>
-                <CompareChipA $productName={product1Chip} />
-              </ChipContainer>
-            )}
-          </TagContainer>
-          {isShow && (
-            <ProductCheckedNameWrap>
-              {similarProducts?.list?.map((product) => (
-                <ProductCheckedNameItem key={product.id} onClick={handleClickSearch}>
-                  {product.name}
-                </ProductCheckedNameItem>
-              ))}
-            </ProductCheckedNameWrap>
-          )}
-        </InputContainer>
-        {/* <InputContainer>
-          <Label>상품2</Label>
-          <TagContainer>
-            <ProductInput
-              value={product2}
-              onChange={handleProductBChange}
-              onKeyDown={handleAddProductB}
-              readOnly={readOnly2}
-            />
-            {isShowChip2 && (
-              <ChipContainer onClick={handleDelete}>
-                <CompareChipB $productName={product2Chip} />
-              </ChipContainer>
-            )}
-          </TagContainer>
-          {isShow2 && (
-            <ProductCheckedNameWrap>
-              {similarProducts2?.list?.map((product) => (
-                <ProductCheckedNameItem key={product.id} onClick={handleClickSearch2}>
-                  {product.name}
-                </ProductCheckedNameItem>
-              ))}
-            </ProductCheckedNameWrap>
-          )}
-        </InputContainer> */}
-        <StyledPrimaryButton type="button" onClick={handleSubmit}>
-          비교하기
-        </StyledPrimaryButton>
+        <form onSubmit={handleSubmit}>
+          <Product1 handleSearchProduct={handleSearchProduct} setSimilarProducts={setSimilarProducts} />
+
+          <Product2 />
+          <StyledPrimaryButton type="submit">비교하기</StyledPrimaryButton>
+        </form>
       </BodyContainer>
       {/* <Loading /> */}
       {/* <CompareTable productOneData={productOneData} /> */}
     </>
   );
 }
-
-const ProductCheckedNameWrap = styled.ul`
-  display: flex;
-  width: 295px;
-  padding: 10px;
-  flex-direction: column;
-  gap: 5px;
-  border-radius: 8px;
-  border: 1px solid var(--color-black-35, #353542);
-  background: var(--color-black-25, #252530);
-  position: absolute;
-  top: 80px;
-  z-index: 1;
-
-  @media (min-width: ${({ theme }) => theme.deviceSizes.tablet}) {
-    width: 360px;
-  }
-`;
-
-const ProductCheckedNameItem = styled.li`
-  display: flex;
-  padding: 6px 20px;
-  gap: 10px;
-  border-radius: 6px;
-  color: var(--color-gray-6e, #6e6e82);
-  ${fontStyle({ w: 400, s: 14, l: 20 })};
-  // ${(props) => props.isMatch && matchedStyles}
-
-  &:hover {
-    background: var(--color-black-35, #353542);
-    color: var(--color-white, #f1f1f5);
-  }
-
-  @media (min-width: ${({ theme }) => theme.deviceSizes.desktop}) {
-    ${fontStyle({ w: 400, s: 16, l: 22 })};
-  }
-`;
